requirements.txt
watchdog
pandas
numpy
Flask
scikit-learn
plotly
psycopg2-binary  # For PostgreSQL
mysql-connector-python  # For MySQL
Flask-SQLAlchemy
Flask-Migrate